requirements.txt
beautifulsoup4==4.12.3
groq==0.11.0
numpy==2.1.1
openai==0.28.0
pandas==2.2.2
pillow==11.0.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.40.2
